Fix fullpage when reloading page #1

+++ src/app/project-detail/project-detail.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Location } from '@angular/common';
 import { MnFullpageService } from 'ng2-fullpage';
+import { AngularFire, FirebaseObjectObservable } from 'angularfire2';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-project-detail',
@@ -8,7 +10,11 @@ import { MnFullpageService } from 'ng2-fullpage';
   styleUrls: ['./project-detail.component.css']
 })
 export class ProjectDetailComponent implements OnInit {
-  constructor(private location: Location, private fullpageService: MnFullpageService) {
+
+  item: FirebaseObjectObservable<any>;
+
+  constructor(private location: Location, private fullpageService: MnFullpageService, af: AngularFire, private router: Router) {
+        this.item = af.database.object('/projects');
    }
 
   goBack(): void {
@@ -16,7 +22,9 @@ export class ProjectDetailComponent implements OnInit {
   }
 
   ngOnInit() {
-    this.fullpageService.destroy('all');
+      if (this.router.navigated) {
+        this.fullpageService.destroy('all');
+      }
   }
 
 }
